requirements.txt
amqp==5.0.1
astroid==2.4.2
billiard==3.6.3.0
braintree==3.45.0
celery==5.0.0
certifi==2020.6.20
amqp==5.0.1
astroid==2.4.2
billiard==3.6.3.0
boto3==1.16.9
botocore==1.19.9
braintree==3.45.0
celery==5.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.16
django-crispy-forms==1.9.2
django-environ==0.4.5
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-storages==1.10.1
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
isort==5.5.4
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumberslite==8.12.10
Pillow==7.2.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
